chore: Modify request header in Coolapk-Scripts.js
This commit modifies the request header in the Coolapk-Scripts.js file. The previous code used `$httpClient.setRequestHeader()` to set the "Accept" header, but it has been updated to use `$request.headers` instead. The modified header is then printed for debugging purposes. This change ensures that the image header issue is fixed and the script properly ends and returns the processed body data.